+++ src/pages/utilities/Converter.tsx
@@ -1,9 +1,9 @@
 import FixedElements from '../../components/FixedElements';
+import { ethers } from 'ethers';
 import { useHistory } from 'react-router-dom';
 import { useRef, useState } from 'react';
-import { ethers } from 'ethers';
 
-export default function NameHash() {
+export default function Converter() {
 	const history = useHistory();
 
 	const hash = useRef(null);
@@ -13,9 +13,12 @@ export default function NameHash() {
 		setError(null);
 		try {
 			if (hash.current.value === '')
-				throw new Error('please enter a dns domain');
-
-			const decoded = ethers.utils.namehash(hash.current.value);
+				throw new Error('please enter a content hash');
+			let decoded: any;
+			if (hash.current.value.startsWith('0x')) {
+				hash.current.value = hash.current.value.slice(2);
+				decoded = ethers.utils.toUtf8String(hash.current.value);
+			} else decoded = ethers.utils.toUtf8Bytes(hash.current.value);
 			setDecoded(decoded);
 		} catch (e) {
 			console.log(e);
@@ -30,21 +33,22 @@ export default function NameHash() {
 				<div className="hero-content text-center text-neutral-content bg-gray-500">
 					<div className="max-w-xl">
 						<h1 className="mb-5 text-5xl font-bold text-black">
-							ENS NameHash Calculator
+							UTF8String / UTF8Bytes Converter
 						</h1>
 						<p className="mb-5 text-black">
-							Please enter an ENS domain to calculate the namehash
+							Please enter either UTF8String or UTF8Bytes
 						</p>
 						<input
 							className="input input-bordered w-full mb-2"
 							ref={hash}
+							onKeyDown={(e) => {
+								if (e.key === 'Enter') decode();
+							}}
 						></input>
 						{error === null ? (
-							<p className="mb-5 text-success mt-2 break-words">{decoded}</p>
+							<p className="mb-5 text-success mt-2">{decoded}</p>
 						) : (
-							<p className="mb-5 text-error mt-2 break-words">
-								{error.message}
-							</p>
+							<p className="mb-5 text-error mt-2">{error.message}</p>
 						)}
 						<button
 							className="btn btn-dark w-full"
@@ -52,12 +56,12 @@ export default function NameHash() {
 								decode();
 							}}
 						>
-							Decode
+							Convert
 						</button>
 						<button
 							className="btn btn-dark w-full mt-2"
 							onClick={() => {
-								history.push('/utilitys/');
+								history.push('/utilities/');
 							}}
 						>
 							Dashboard
